backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1
Flask-Migrate==4.1.0
python-dotenv==1.1.1
bcrypt==4.3.0
psycopg2-binary==2.9.10
reportlab==4.0.7
